+++ tests/testthat/test-set_endpoint.R
@@ -21,3 +21,29 @@ test_that("resets response format correctly according to reset_format argument",
 	expect_equal(q3$body$format, "geojson")
 	expect_null(q4$body$format)
 })
+
+test_that("returns unmodified endpoint if grouping arg is missing", {
+	q1 <- ohsome_query("elements/count/groupBy/boundary/groupBy/tag")
+	q2 <- set_grouping(q1)
+	expect_equal(httr::parse_url(q1$url)$path, httr::parse_url(q2$url)$path)
+	
+	q1 <- ohsome_query("elements/count")
+	q2 <- set_grouping(q1)
+	expect_equal(httr::parse_url(q1$url)$path, httr::parse_url(q2$url)$path)
+})
+
+test_that("correctly modifies grouping", {
+	q1 <- ohsome_query("elements/count/groupBy/boundary")
+	q2 <- set_grouping(q1, "tag")
+	expect_equal(httr::parse_url(q2$url)$path, "v1/elements/count/groupBy/tag")
+	
+	q1 <- ohsome_query("elements/count")
+	q2 <- set_grouping(q1, "tag")
+	expect_equal(httr::parse_url(q2$url)$path, "v1/elements/count/groupBy/tag")
+})
+
+test_that("removes grouping if grouping = NULL", {
+	q1 <- ohsome_query("elements/count/groupBy/boundary")
+	q2 <- set_grouping(q1, grouping = NULL)
+	expect_equal(httr::parse_url(q2$url)$path, "v1/elements/count")
+})